more layout things
decided against trying to experiment with A-star. Godot has
implementation for 2D A-star already. Priority Queues and Heaps are
included by the C# standard library.

Godot has support for automatically generating nav-meshes, which will
replace the A-star pathfinding idea we had. this makes the pathfinding
thing super simple but infeasible to try to do on a terminal. need to
experiment with Godot and learn how it works in order to start progress
on pathfinding with a nav-mesh.

did a lot of laying out for various classes that will probably be
useful. worked on the turn based combat handler, which will take a list
of all characters in a fight and convert them into "fighters". terminal
version will have no map.